Cleaned up the regridding function.
